Update jackpot amount format. Update command and query classes to records.

+++ VirtualRoulette.Application/DTOs/MakeBetResponseDto.cs
@@ -12,7 +12,7 @@ namespace VirtualRoulette.Application.DTOs
         public BetStatus Status { get; set; }
         public Guid? SpinId { get; set; }
         public int? WinningNumber { get; set; }
-        public int? WonAmount { get; set; }
+        public long? WonAmount { get; set; }
         public string? Message { get; set; }
     }
 }
